requirements.txt
customtkinter
selenium
webdriver-manager
beautifulsoup4
pandas
openpyxl
requests
Pillow
python-dotenv
appdirs
